fixed crashing app due to null values
